fix: resolve type-safety debt and mypy errors
Resolves 144 mypy errors across 26 files by:
- Fixing broken imports (FRAMEWORK_VERSION).
- Adding explicit casts and conversions for SDK returns.
- Adding missing type annotations and resolving redefinitions.
- Handling optional dependencies for static analysis.
- Typing complex dictionaries and lists properly.

+++ cognee/memory/hierarchical/decay.py
@@ -22,4 +22,4 @@ class MemoryDecayModel:
             memory.get("emotional_valence", 0) * self.genome.emotional_weight_multiplier
         )
 
-        return min(1.0, base * usage * emotional_boost)
+        return float(min(1.0, base * usage * emotional_boost))
